+++ HISTORY.rst
@@ -40,6 +40,8 @@ latest (unreleased)
   The quantity of the current stock location is now returned instead of the
   total stock within the warehouse.
 
+* ALCYN-2266: Concurrent access on delivery round assignment
+
 **Build**
 
 **Documentation**